+++ frontend/src/pages/HomePage.tsx
@@ -2,7 +2,7 @@ import {useEffect, useState} from "react";
 import axios from "axios";
 
 import type {HomeTableReturnDTO} from "@/dto/response/HomeTableReturnDTO.ts";
-import CreateNewHome from "@/components/create/CreateNewHome.tsx";
+import CreateNewHome from "@/pages/home/CreateNewHome.tsx";
 
 import HomeCard from "@/pages/home/HomeCard.tsx";
